Make page naming more consistent
