+++ frontend/src/component/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo, JSX } from 'react';
 import { RouteProps } from 'react-router';
 import { useCurrentUserCtx } from '../contexts/CurrentUserCtx';
-import { Login } from '../page/Login';
+import { LoginPage } from '../page/LoginPage';
 
 export interface PrivateRouteProps {
     children: JSX.Element;
@@ -16,5 +16,5 @@ export const PrivateRoute = ({
     const canView = useMemo(() => {
         return currentUser && currentUser.permission_level >= permission;
     }, [currentUser, permission]);
-    return canView ? children : <Login />;
+    return canView ? children : <LoginPage />;
 };
