+++ frontend/src/page/AdminImportPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Typography from '@mui/material/Typography';
 import Grid from '@mui/material/Unstable_Grid2';
 
-export const AdminImport = () => {
+export const AdminImportPage = () => {
     return (
         <Grid container>
             <Grid xs>
